\name{ACL394} 
\docType{data} 
\alias{ACL394} 
\alias{ACL394filtered} 
\alias{pos1.cell.counter}
\alias{pdata}
\alias{X} 
\title{Time Dependent Dose Response of Yeast Cells to Mating Pheromone} 
\description{
This datasets was generated by Cell-ID, from an experiment done in 2004 by Alejandro Colman-Lerner and Andrew Gordon at the Molecular Science Institute (MSI). \emph{Saccharomices cerevisiae} yeast cells of strain TCY3154 (MATa, bar1, prm1::Pprm1-YFP::HIS+, trp1::Pact1-CFP::TRP1) where stimulated with different doses of alpha-factor pheromone 10 minutes before the first time point. Images where adquired every 15 minutes for 3.5 hours. In the dataset there are 3 positions per treatment. The strain TCY3154 was derived from ACL394, a W303 derivative. 

data(ACL394) loads the unfiltered dataset, while data(ACL394filtered) loads the dataset with filters applied.

pos1.cell.counter is a data.frame as returned by ImageJ's plugin Cell Counter.

pdata is a data.frame with the description of each position.

} 
\usage{X} 
\format{a cell.data object} 
\references{
Colman-Lerner et al. (2005). "Regulated cell-to-cell variation in a cell-fate decision system." Nature 437(7059):699-706.
}
\keyword{datasets}