\name{RcellData-package}
\docType{package}
\alias{RcellData-package}
\alias{RcellData}
\title{
  Example Dataset for Package Rcell
}
\description{
 Example Cell-ID dataset for Rcell package. Includes images and loaded cell data object. 
 }
\details{
\tabular{ll}{
Package: \tab Rcell\cr
Type: \tab Package\cr
Version: \tab 1.3-1 \cr
Date: \tab 2015-05-28 \cr
License: \tab GPL-2\cr
}
}

\author{
Alan Bush
Maintainer: Alan Bush <abush@fbmc.fcen.uba.ar>
}
\references{
\url{http://sourceforge.net/projects/cell-id}
}
\seealso{
  \pkg{Rcell}
}
\keyword{package}
